+++ zed_components/src/zed_camera/include/zed_camera_component.hpp
@@ -433,6 +433,7 @@ private:
   bool _debugVideoDepth = false;
   bool _debugCamCtrl = false;
   bool _debugPointCloud = false;
+  bool _debugTf = false;
   bool _debugPosTracking = false;
   bool _debugGnss = false;
   bool _debugSensors = false;